week_2/airflow/requirements.txt
apache-airflow-providers-google
pyarrow
pandas
sqlalchemy
psycopg2-binary